Fixed size_t

+++ src/io/endianness.hpp
@@ -5,6 +5,8 @@
 # include "config.hpp"
 #endif
 
+#include <unistd.h>
+
 #include <metaprogramming/inline.hpp>
 
 namespace nissa
@@ -60,7 +62,7 @@ namespace nissa
     
     /// Fills the output
     ByteSwapper out;
-    for(int i=0;i<sizeof(T);i++)
+    for(size_t i=0;i<sizeof(T);i++)
       out.c[i]=in.c[sizeof(T)-1-i];
     
     return out.t;
